Add Docker/Compose setup, verified end to end
- Multi-stage Dockerfile: golang:1.25-alpine builder, distroless
  (nonroot) runtime image; coupons.idx copied in as a committed
  build artifact rather than rebuilt in-image
- docker-compose.yml wiring postgres (with healthcheck) + backend,
  migrations run on boot
- Fixes found by actually running it: go.mod's version directive had
  drifted to the local toolchain's 1.27.1 via `go mod tidy`, pinned
  back down to the real transitive minimum (1.25); product listing
  was sorting "1, 10, 2, ..." because id is TEXT, now ORDER BY id::int;
  request duration was logging as raw nanoseconds under slog's JSON
  handler, switched to a duration_ms field
- Full curl pass against a clean `docker compose up --build`: health/
  readiness, product list/get (200/400/404), and the order flow
  (401/403/422 x3/200) all verified, including HAPPYHRS/SUPER100
  against the real coupon index, with the order persisted correctly
  across the orders/order_items transaction

+++ internal/product/postgres_repository.go
@@ -19,7 +19,10 @@ func NewPostgresRepository(pool *pgxpool.Pool) *PostgresRepository {
 }
 
 func (r *PostgresRepository) List(ctx context.Context) ([]Product, error) {
-	rows, err := r.pool.Query(ctx, `SELECT id, name, price::float8, category FROM products ORDER BY id`)
+	// id is TEXT for storage-backend flexibility, but holds numeric
+	// values today — order numerically so listings read naturally
+	// rather than "1, 10, 2, 3, ...".
+	rows, err := r.pool.Query(ctx, `SELECT id, name, price::float8, category FROM products ORDER BY id::int`)
 	if err != nil {
 		return nil, err
 	}
